requirements.txt
vaderSentiment
textblob
google-cloud-texttospeech